+++ gems/quilt_rails/test/test_helper.rb
@@ -3,8 +3,6 @@
 # Configure Rails Environment
 ENV["RAILS_ENV"] = "test"
 
-require_relative "../test/dummy/config/environment"
-
 require 'minitest/autorun'
 require 'rails'
 require 'mocha/minitest'