+++ packages/react-bugsnag/src/types.ts
@@ -1,0 +1,3 @@
+import {Client} from '@bugsnag/js';
+
+export type ErrorLogger = Pick<Client, 'notify' | 'leaveBreadcrumb'>;