+++ packages/react-server/src/webpack-plugin/index.ts
@@ -1,0 +1,1 @@
+export {ReactServerPlugin} from './webpack-plugin';