+++ packages/jest-mock-apollo/CHANGELOG.md
@@ -7,6 +7,12 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [5.0.0] - 2020-06-24
+
+### Fixed
+
+- Mock client now throws useful errors and fails tests when required mocks are not provided
+
 ## [4.0.6] - 2020-03-13
 
 - Update dev dependencies from `react-apollo` to `@apollo/react-common` and `@apollo/react-hoc` ([#1321](https://github.com/Shopify/quilt/pull/1321))